requirements.txt
# Core dependencies
streamlit==1.45.1
python-dotenv==1.1.0

# LangChain ecosystem
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langchain-google-community==2.0.7

# LangGraph
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-checkpoint-sqlite==2.0.10
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42

# OpenAI
openai==1.82.0

# Google APIs
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
googlemaps==4.10.0

# Data processing
pandas==2.2.3
numpy==2.2.6
pydantic==2.11.5
pydantic-settings==2.9.1

# HTTP requests
requests==2.32.3
httpx==0.28.1
aiohttp==3.11.18

# Other utilities
python-dateutil
PyYAML==6.0.2
tiktoken==0.9.0
tenacity==9.1.2
tqdm==4.67.1
